Code refactoring, optimization

+++ src/main/java/com/bsu/lab/AccountingSystem/dao/EntranceRepository.java
@@ -4,7 +4,7 @@ import com.bsu.lab.AccountingSystem.domain.Entrance;
 import com.bsu.lab.AccountingSystem.domain.Floor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EntranceRepository extends JpaRepository<Entrance, Integer> {
+public interface EntranceRepository extends JpaRepository<Entrance, Long> {
 
     Entrance getByFloorsContains(Floor floor);
 }
