+++ src/main/java/com/bsu/lab/AccountingSystem/dao/FloorRepository.java
@@ -5,7 +5,7 @@ import com.bsu.lab.AccountingSystem.domain.Floor;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface FloorRepository extends JpaRepository<Floor, Integer> {
+public interface FloorRepository extends JpaRepository<Floor, Long> {
 
     Floor getByFlatsContains(Flat flat);
 }
